REFACTOR: move buttons widgets to buttons folder
